refactor: 상품 검색/필터 헤더 Sticky 노출 로직 훅으로 분리
