+++ src/pages/home/hooks/useProductStickyHeader.ts
@@ -1,0 +1,55 @@
+import { useEffect, useRef, useState } from 'react';
+
+interface UseProductStickyHeaderParams {
+  searchBarRef: React.RefObject<HTMLDivElement | null>;
+  filterListRef: React.RefObject<HTMLDivElement | null>;
+}
+
+export const useProductStickyHeader = ({
+  searchBarRef,
+  filterListRef,
+}: UseProductStickyHeaderParams) => {
+  const lastScrollYRef = useRef(0);
+  const [isFilterSticky, setIsFilterSticky] = useState(false);
+  const [showStickySearchBar, setShowStickySearchBar] = useState(false);
+
+  useEffect(() => {
+    // handleScroll: 스크롤 위치/방향에 따라 sticky 상태와 검색바 노출 상태 갱신
+    const handleScroll = () => {
+      if (!searchBarRef.current || !filterListRef.current) return;
+
+      const currentY = window.scrollY || window.pageYOffset;
+      const isScrollUp = currentY < lastScrollYRef.current;
+      lastScrollYRef.current = currentY;
+
+      const filterTop = filterListRef.current.getBoundingClientRect().top;
+      const searchBarTop = searchBarRef.current.getBoundingClientRect().top;
+
+      // 필터칩 행 상단이 viewport 상단에 닿으면 sticky 시작
+      if (!isFilterSticky && filterTop <= 0) {
+        setIsFilterSticky(true);
+      }
+
+      if (isFilterSticky) {
+        // 스크롤 업 중 원래 검색바 상단이 다시 화면 안으로 들어오면 sticky 해제
+        if (isScrollUp && searchBarTop >= 0) {
+          setIsFilterSticky(false);
+          setShowStickySearchBar(false);
+          return;
+        }
+        // sticky 상태에서는 스크롤 업일 때만 검색바 노출
+        setShowStickySearchBar(isScrollUp);
+      }
+    };
+
+    window.addEventListener('scroll', handleScroll, { passive: true });
+    return () => {
+      window.removeEventListener('scroll', handleScroll);
+    };
+  }, [filterListRef, isFilterSticky, searchBarRef]);
+
+  return {
+    isFilterSticky,
+    showStickySearchBar,
+  };
+};
